review: fix the findings from the PR review
dichotomy: LoopScan() read opt_quants[] uninitialized when the first pass
broke out on a failed allocation. Guard the post-search block with ok_; the
bitstream writes keep their inner guard, since StoreRunLevels() can still
clear ok_.

enc: 'W * H + 2 * uv_w * uv_h' wrapped at 65535x65535 to a plausible 2.1GB,
so the allocation could succeed and the conversion write past the end. Size
the planes as size_t, and cast the six 'width * height / 4' sink hints and
the SjpegRiskiness() pixel count the same way.

includes: <assert.h>, <string.h>, <string>, bit_writer.h, <utility>, each
where it is used rather than reached through sjpegi.h.

tests: Threads (TSAN race on the lazy tables), TargetSize (1.4% convergence,
against 4.8-8.4% when YUV_400 is charged for a second quantization matrix),
NegativeStrides, and odd 17x13 dimensions in the stride tests. Each one
verified to fail without the commit it covers.

+++ src/jpeg_tools.cc
@@ -19,6 +19,7 @@
 #include <string.h>   // for memset
 
 #include <cstdlib>
+#include <utility>   // for std::swap
 #include <vector>
 
 #include "sjpegi.h"
@@ -212,7 +213,7 @@ SjpegYUVMode SjpegRiskiness(const uint8_t* rgb,
   if (num_samples > 0.) gray_count /= num_samples;
 
   // number of pixels evaluated
-  const double frac = 100. * count / (width * height);
+  const double frac = 100. * count / ((double)width * height);
   // if less than 1% of pixels were evaluated -> below noise level.
   if (frac < 1.) total_score = 0.;
